machine learning/requirements.txt
numpy
pandas
Pillow
opencv-python
torch
torchvision
pendulum
segmentation-models-pytorch
albumentations